refactor(core): extract stop orchestration from daemon god file
Second cut on daemon (still a god file after runner_files, #9417). Extract
the daemon-stop orchestration cluster into stop.rs: the public entry points
(stop, stop_with_force, stop_for_lease, force_stop_for_lease), their
unlocked/lease-scoped internals (stop_unlocked*, stop_with_force_for_lease,
force_stop_for_lease_unlocked, reconcile_absent_lease_stop), and the
active-job stop-guard (active_daemon_job_ids,
active_jobs_block_daemon_stop_error).

stop.rs uses 'use super::*' to inherit the parent's state types, operation
lock, and persistence helpers; a few private mod.rs items the cluster needs
(state_path, acquire_daemon_operation_lock, DaemonLeaseIdentity,
DAEMON_STARTUP_TOKEN_ENV, FORCE_STOP_WAIT) are now pub(super). Public entry
points re-exported (daemon::stop etc. unchanged); the internal fns mod.rs /
control.rs still call are re-exported pub(crate)/imported. External contract
unchanged.

Behavior-preserving: daemon/mod.rs 3195 -> 2855 lines, stop.rs 358. All
stop-specific tests pass. (Same 3 pre-existing order-dependent flaky daemon
tests as #9417 — 130 passed / 3 failed, identical on clean main.)
